# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API Documentation",
        "title": "Synapsis E-Commerce",
        "contact": {
            "name": "Armada Muhammad Siswanto",
            "email": "armadamuhammads@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/products/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a Product by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Get a Product by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product data",
                        "schema": {
                            "$ref": "#/definitions/model.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update Product by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Update Product by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Product data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ProductAPI"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product data",
                        "schema": {
                            "$ref": "#/definitions/model.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete Product by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Delete Product by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    }
                }
            }
        },
...
        "/users/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a User by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get a User by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User data",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update User by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update User by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserAPI"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User data",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "lib.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "response message",
                    "type": "string"
                },
                "status": {
                    "description": "http status",
                    "type": "integer"
                }
            }
        },
...
        "model.Product": {
            "type": "object",
            "properties": {
                "additional": {
                    "type": "string"
                },
                "categoryid": {
                    "description": "CategoryID",
                    "type": "string",
                    "format": "uuid"
                },
                "description": {
                    "description": "Description",
                    "type": "string"
                },
                "name": {
                    "description": "Name",
                    "type": "string",
                    "example": "Sepatu"
                },
                "price": {
                    "description": "Price",
                    "type": "number",
                    "example": 10000
                },
                "sku": {
                    "description": "SKU",
                    "type": "string",
                    "example": "PR-0000001"
                }
            }
        },
        "model.ProductAPI": {
            "type": "object",
            "properties": {
                "categoryid": {
                    "description": "CategoryID",
                    "type": "string",
                    "format": "uuid"
                },
                "description": {
                    "description": "Description",
                    "type": "string"
                },
                "name": {
                    "description": "Name",
                    "type": "string",
                    "example": "Sepatu"
                },
                "price": {
                    "description": "Price",
                    "type": "number",
                    "example": 10000
                },
                "sku": {
                    "description": "SKU",
                    "type": "string",
                    "example": "PR-0000001"
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "properties": {
                "additional": {
                    "type": "string"
                },
                "email": {
                    "description": "Email",
                    "type": "string",
                    "example": "armadamuhammads@gmail.com"
                },
                "firstname": {
                    "description": "FirstName",
                    "type": "string",
                    "example": "Armada"
                },
                "isadmin": {
                    "description": "IsAdmin",
                    "type": "boolean"
                },
                "isverified": {
                    "description": "IsVerified",
                    "type": "boolean"
                },
                "lastlogin": {
                    "description": "LastLogin",
                    "type": "string",
                    "format": "date-time"
                },
                "lastname": {
                    "description": "LastName",
                    "type": "string",
                    "example": "Muhammad Siswanto"
                },
                "password": {
                    "description": "Password",
                    "type": "string"
                },
                "phonenumber": {
                    "description": "PhoneNumber",
                    "type": "string",
                    "example": "089678009400"
                },
                "username": {
                    "description": "Username",
                    "type": "string"
                },
                "verificationcode": {
                    "description": "VerificationCode",
                    "type": "string",
                    "example": "2024"
                }
            }
        },
        "model.UserAPI": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "Email",
                    "type": "string",
                    "example": "armadamuhammads@gmail.com"
                },
                "firstname": {
                    "description": "FirstName",
                    "type": "string",
                    "example": "Armada"
                },
                "isadmin": {
                    "description": "IsAdmin",
                    "type": "boolean"
                },
                "isverified": {
                    "description": "IsVerified",
                    "type": "boolean"
                },
                "lastlogin": {
                    "description": "LastLogin",
                    "type": "string",
                    "format": "date-time"
                },
                "lastname": {
                    "description": "LastName",
                    "type": "string",
                    "example": "Muhammad Siswanto"
                },
                "password": {
                    "description": "Password",
                    "type": "string"
                },
                "phonenumber": {
                    "description": "PhoneNumber",
                    "type": "string",
                    "example": "089678009400"
                },
                "username": {
                    "description": "Username",
                    "type": "string"
                },
                "verificationcode": {
                    "description": "VerificationCode",
                    "type": "string",
                    "example": "2024"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "TokenKey": {
            "type": "apiKey",
            "name": "x-Token",
            "in": "header"
        }
    }
}